requirements-vercel.txt
fastapi>=0.104.1
uvicorn>=0.24.0
requests>=2.31.0
python-multipart>=0.0.6
jinja2>=3.1.2
python-dotenv>=1.0.0
pillow>=10.0.0
aiofiles>=23.2.1
pydantic>=2.5.0
httpx>=0.25.2
openai>=1.58.0
